#5 Was been extracted BaseItIsAnyParameterProvider
